backend/requirements.txt
fastapi[all]==0.115.0
uvicorn==0.32.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
asyncpg==0.30.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
celery==5.4.0
redis==5.1.1
httpx==0.27.2
google-generativeai==0.8.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0
